pyproject.toml
[project]
name = "perspective-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.185.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
